cloud-functions/google_login/requirements.txt
Flask==3.0.3
PyJWT==2.8.0
PyMongo==4.7.3
google-auth==2.29.0
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-storage==2.16.0
requests==2.32.3